+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/endpoint/PollsApi.kt
@@ -17,50 +17,169 @@
 package io.getstream.chat.android.client.api2.endpoint
 
 import io.getstream.chat.android.client.api.AuthenticatedApi
-import io.getstream.chat.android.client.api2.model.requests.PollRequest
-import io.getstream.chat.android.client.api2.model.requests.PollUpdateRequest
+import io.getstream.chat.android.client.api2.model.requests.CreatePollRequest
+import io.getstream.chat.android.client.api2.model.requests.PartialUpdatePollRequest
 import io.getstream.chat.android.client.api2.model.requests.PollVoteRequest
-import io.getstream.chat.android.client.api2.model.requests.SuggestPollOptionRequest
+import io.getstream.chat.android.client.api2.model.requests.QueryPollVotesRequest
+import io.getstream.chat.android.client.api2.model.requests.QueryPollsRequest
+import io.getstream.chat.android.client.api2.model.requests.UpdatePollRequest
+import io.getstream.chat.android.client.api2.model.requests.UpstreamOptionDto
 import io.getstream.chat.android.client.api2.model.response.CompletableResponse
+import io.getstream.chat.android.client.api2.model.response.PollOptionResponse
 import io.getstream.chat.android.client.api2.model.response.PollResponse
 import io.getstream.chat.android.client.api2.model.response.PollVoteResponse
-import io.getstream.chat.android.client.api2.model.response.SuggestPollOptionResponse
+import io.getstream.chat.android.client.api2.model.response.QueryPollVotesResponse
+import io.getstream.chat.android.client.api2.model.response.QueryPollsResponse
 import io.getstream.chat.android.client.call.RetrofitCall
 import retrofit2.http.Body
 import retrofit2.http.DELETE
+import retrofit2.http.GET
 import retrofit2.http.PATCH
 import retrofit2.http.POST
+import retrofit2.http.PUT
 import retrofit2.http.Path
 
+/**
+ * Retrofit interface defining the Polls API.
+ */
+@Suppress("TooManyFunctions")
 @AuthenticatedApi
 internal interface PollsApi {
 
     /**
      * Creates a new poll.
      *
-     * @param pollRequest The poll request.
+     * See [Create poll](https://getstream.github.io/protocol/#/product%3Achat/CreatePoll).
+     *
+     * @param body Body holding the poll properties to be created.
      *
      * @return The poll response.
      */
     @POST("/polls")
-    fun createPoll(@Body pollRequest: PollRequest): RetrofitCall<PollResponse>
+    fun createPoll(@Body body: CreatePollRequest): RetrofitCall<PollResponse>
 
     /**
-     * Suggest a new option for a poll.
+     * Updates an existing poll.
+     *
+     * See [Update poll](https://getstream.github.io/protocol/#/product%3Achat/UpdatePoll).
+     *
+     * @param body Body holding the poll properties to be updated.
+     */
+    @PUT("/polls")
+    fun updatePoll(@Body body: UpdatePollRequest): RetrofitCall<PollResponse>
+
+    /**
+     * Deletes a poll.
+     *
+     * See [Delete poll](https://getstream.github.io/protocol/#/product%3Achat/DeletePoll).
+     *
+     * @param pollId The ID of the poll to delete.
+     */
+    @DELETE("/polls/{poll_id}")
+    fun deletePoll(@Path("poll_id") pollId: String): RetrofitCall<CompletableResponse>
+
+    /**
+     * Retrieves a poll by its ID.
+     *
+     * See [Get poll](https://getstream.github.io/protocol/#/product%3Achat/GetPoll).
      *
      * @param pollId The poll ID.
-     * @param suggestPollOptionRequest The suggest poll option request.
+     *
+     * @return The poll response.
+     */
+    @GET("/polls/{poll_id}")
+    fun getPoll(@Path("poll_id") pollId: String): RetrofitCall<PollResponse>
+
+    /**
+     * Updates a poll.
+     *
+     * See [Partial update poll](https://getstream.github.io/protocol/#/product%3Achat/UpdatePollPartial).
+     *
+     * @param pollId The poll ID.
+     * @param body The poll update request.
+     *
+     * @return The poll response.
+     */
+    @PATCH("/polls/{poll_id}")
+    fun partialUpdatePoll(
+        @Path("poll_id") pollId: String,
+        @Body body: PartialUpdatePollRequest,
+    ): RetrofitCall<PollResponse>
+
+    /**
+     * Create a new option for a poll.
+     *
+     * See [Create poll option](https://getstream.github.io/protocol/#/product%3Achat/CreatePollOption).
+     *
+     * @param pollId The poll ID.
+     * @param body The create poll option request.
      *
      * @return The suggest poll option response.
      */
     @POST("/polls/{poll_id}/options")
-    fun suggestPollOption(
+    fun createPollOption(
         @Path("poll_id") pollId: String,
-        @Body suggestPollOptionRequest: SuggestPollOptionRequest,
-    ): RetrofitCall<SuggestPollOptionResponse>
+        @Body body: UpstreamOptionDto,
+    ): RetrofitCall<PollOptionResponse>
+
+    /**
+     * Update an option for a poll.
+     *
+     * See [Update poll option](https://getstream.github.io/protocol/#/product%3Achat/UpdatePollOption).
+     *
+     * @param pollId The poll ID.
+     * @param body The update poll option request.
+     */
+    @PUT("/polls/{poll_id}/options")
+    fun updatePollOption(
+        @Path("poll_id") pollId: String,
+        @Body body: UpstreamOptionDto,
+    ): RetrofitCall<PollOptionResponse>
+
+    /**
+     * Deletes an option from a poll.
+     *
+     * See [Delete poll option](https://getstream.github.io/protocol/#/product%3Achat/DeletePollOption).
+     *
+     * @param pollId The poll ID.
+     * @param optionId The option ID.
+     */
+    @DELETE("/polls/{poll_id}/options/{option_id}")
+    fun deletePollOption(
+        @Path("poll_id") pollId: String,
+        @Path("option_id") optionId: String,
+    ): RetrofitCall<CompletableResponse>
+
+    // MISSING: Get poll option
+
+    /**
+     * Queries votes for a specific poll based on the provided criteria.
+     *
+     * See: [Query poll votes](https://getstream.github.io/protocol/#/product%3Achat/QueryPollVotes).
+     *
+     * @param pollId The poll ID.
+     * @param body The query poll votes request.
+     */
+    @POST("/polls/{poll_id}/votes")
+    fun queryPollVotes(
+        @Path("poll_id") pollId: String,
+        @Body body: QueryPollVotesRequest,
+    ): RetrofitCall<QueryPollVotesResponse>
+
+    /**
+     * Queries polls based on the provided criteria.
+     *
+     * See: [Query polls](https://getstream.github.io/protocol/#/product%3Achat/QueryPolls).
+     *
+     * @param body The query polls request.
+     */
+    @POST("/polls/query")
+    fun queryPolls(@Body body: QueryPollsRequest): RetrofitCall<QueryPollsResponse>
 
     /**
      * Casts a vote on a poll.
+     *
+     * See: [Cast poll vote](https://getstream.github.io/protocol/#/product%3Achat/CastPollVote).
      *
      * @param messageId The message ID.
      * @param pollId The poll ID.
@@ -76,7 +195,9 @@ internal interface PollsApi {
     ): RetrofitCall<PollVoteResponse>
 
     /**
-     * Removes a vote on a poll.
+     * Deletes a vote on a poll.
+     *
+     * See: [Delete poll vote](https://getstream.github.io/protocol/#/product%3Achat/DeletePollVote).
      *
      * @param messageId The message ID.
      * @param pollId The poll ID.
@@ -90,26 +211,4 @@ internal interface PollsApi {
         @Path("poll_id") pollId: String,
         @Path("vote_id") voteId: String,
     ): RetrofitCall<PollVoteResponse>
-
-    /**
-     * Updates a poll.
-     *
-     * @param pollId The poll ID.
-     * @param pollUpdateRequest The poll update request.
-     *
-     * @return The poll response.
-     */
-    @PATCH("/polls/{poll_id}")
-    fun updatePoll(
-        @Path("poll_id") pollId: String,
-        @Body pollUpdateRequest: PollUpdateRequest,
-    ): RetrofitCall<PollResponse>
-
-    /**
-     * Deletes a poll.
-     *
-     * @param pollId The ID of the poll to delete.
-     */
-    @DELETE("/polls/{poll_id}")
-    fun deletePoll(@Path("poll_id") pollId: String): RetrofitCall<CompletableResponse>
 }