+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/api2/MoshiChatApiTestArguments.kt
@@ -42,6 +42,7 @@ import io.getstream.chat.android.client.api2.model.response.FlagResponse
 import io.getstream.chat.android.client.api2.model.response.MessageResponse
 import io.getstream.chat.android.client.api2.model.response.MessagesResponse
 import io.getstream.chat.android.client.api2.model.response.MuteUserResponse
+import io.getstream.chat.android.client.api2.model.response.PollOptionResponse
 import io.getstream.chat.android.client.api2.model.response.PollResponse
 import io.getstream.chat.android.client.api2.model.response.PollVoteResponse
 import io.getstream.chat.android.client.api2.model.response.QueryBannedUsersResponse
@@ -49,12 +50,13 @@ import io.getstream.chat.android.client.api2.model.response.QueryBlockedUsersRes
 import io.getstream.chat.android.client.api2.model.response.QueryChannelsResponse
 import io.getstream.chat.android.client.api2.model.response.QueryDraftMessagesResponse
 import io.getstream.chat.android.client.api2.model.response.QueryMembersResponse
+import io.getstream.chat.android.client.api2.model.response.QueryPollVotesResponse
+import io.getstream.chat.android.client.api2.model.response.QueryPollsResponse
 import io.getstream.chat.android.client.api2.model.response.QueryThreadsResponse
 import io.getstream.chat.android.client.api2.model.response.ReactionResponse
 import io.getstream.chat.android.client.api2.model.response.ReactionsResponse
 import io.getstream.chat.android.client.api2.model.response.ReminderResponse
 import io.getstream.chat.android.client.api2.model.response.SearchMessagesResponse
-import io.getstream.chat.android.client.api2.model.response.SuggestPollOptionResponse
 import io.getstream.chat.android.client.api2.model.response.SyncHistoryResponse
 import io.getstream.chat.android.client.api2.model.response.ThreadResponse
 import io.getstream.chat.android.client.api2.model.response.TokenResponse
@@ -81,7 +83,7 @@ import io.getstream.result.Result
 import okhttp3.ResponseBody
 import org.junit.jupiter.params.provider.Arguments
 
-@Suppress("UNUSED")
+@Suppress("UNUSED", "LargeClass")
 internal object MoshiChatApiTestArguments {
 
     @JvmStatic
@@ -504,27 +506,55 @@ internal object MoshiChatApiTestArguments {
     fun removePollVoteInput() = pollVoteResponseArguments()
 
     @JvmStatic
+    fun partialUpdatePollInput() = pollResponseArguments()
+
+    @JvmStatic
     fun closePollInput() = pollResponseArguments()
 
     @JvmStatic
     fun createPollInput() = pollResponseArguments()
 
     @JvmStatic
+    fun updatePollInput() = pollResponseArguments()
+
+    @JvmStatic
+    fun getPollInput() = pollResponseArguments()
+
+    @JvmStatic
     fun deletePollInput() = completableResponseArguments()
 
     @JvmStatic
-    fun suggestPollOptionInput() = listOf(
+    fun queryPollsInput() = listOf(
         Arguments.of(
-            RetroSuccess(
-                SuggestPollOptionResponse(
-                    duration = randomString(),
-                    poll_option = Mother.randomDownstreamOptionDto(),
-                ),
-            ).toRetrofitCall(),
+            RetroSuccess(Mother.randomQueryPollsResponse()).toRetrofitCall(),
             Result.Success::class,
         ),
-        Arguments.of(RetroError<SuggestPollOptionResponse>(statusCode = 500).toRetrofitCall(), Result.Failure::class),
+        Arguments.of(
+            RetroError<QueryPollsResponse>(statusCode = 500).toRetrofitCall(),
+            Result.Failure::class,
+        ),
     )
+
+    @JvmStatic
+    fun queryPollVotesInput() = listOf(
+        Arguments.of(
+            RetroSuccess(Mother.randomQueryPollVotesResponse()).toRetrofitCall(),
+            Result.Success::class,
+        ),
+        Arguments.of(
+            RetroError<QueryPollVotesResponse>(statusCode = 500).toRetrofitCall(),
+            Result.Failure::class,
+        ),
+    )
+
+    @JvmStatic
+    fun createPollOptionInput() = pollOptionResponseArguments()
+
+    @JvmStatic
+    fun updatePollOptionInput() = pollOptionResponseArguments()
+
+    @JvmStatic
+    fun deletePollOptionInput() = completableResponseArguments()
 
     @JvmStatic
     fun createReminderInput() = reminderResponseArguments()
@@ -759,6 +789,19 @@ internal object MoshiChatApiTestArguments {
             Result.Success::class,
         ),
         Arguments.of(RetroError<PollResponse>(statusCode = 500).toRetrofitCall(), Result.Failure::class),
+    )
+
+    private fun pollOptionResponseArguments() = listOf(
+        Arguments.of(
+            RetroSuccess(
+                PollOptionResponse(
+                    duration = randomString(),
+                    poll_option = Mother.randomDownstreamOptionDto(),
+                ),
+            ).toRetrofitCall(),
+            Result.Success::class,
+        ),
+        Arguments.of(RetroError<PollOptionResponse>(statusCode = 500).toRetrofitCall(), Result.Failure::class),
     )
 
     private fun pollVoteResponseArguments() = listOf(