+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/list/internal/poll/AllPollOptionsDialogFragment.kt
@@ -210,8 +210,8 @@ public class AllPollOptionsDialogFragment : AppCompatDialogFragment() {
                 pollState.value
                     ?.let { poll ->
                         poll.ownVotes.firstOrNull { it.optionId == option.id }
-                            ?.let { chatClient.removePollVote(messageId, poll.id, it) }
-                            ?: chatClient.castPollVote(messageId, poll.id, option)
+                            ?.let { chatClient.removePollVote(messageId, poll.id, it.id) }
+                            ?: chatClient.castPollVote(messageId, poll.id, option.id)
                     }?.await()
             }
         }